// src/app/enums/process.enum.ts
export enum ProcessEnum {
    Inpection = 'Inpection',
    Planes = 'Planes',
    SiteTest = 'SiteTest',
    DailyReport = 'DailyReport',
    AssignLocsToUser = 'AssignLocsToUser',
    CheckHistInspLoc = 'CheckHistInspLoc',
    CheckHistInspOper = 'CheckHistInspOper',
    CustomerCheck = 'CustomerCheck',
    CheckHistReportLoc = 'CheckHistReportLoc',
    CheckHistReportOper = 'CheckHistReportOper',
}